Rest API finished.

+++ Project33/server/controllers/feed.js
@@ -1,28 +1,34 @@
-const { validationResult } = require("express-validator");
-const Post = require("../models/post");
-const fs = require("fs");
-const path = require("path");
-const { clear } = require("console");
+const fs = require('fs');
+const path = require('path');
+
+const { validationResult } = require('express-validator/check');
+
+const Post = require('../models/post');
+const User = require('../models/user');
 
 exports.getPosts = (req, res, next) => {
   const currentPage = req.query.page || 1;
-  const perPage = 2; // 2 posts per page
+  const perPage = 2;
   let totalItems;
-  Post.find().countDocuments()
-    .then((count) => {
+  Post.find()
+    .countDocuments()
+    .then(count => {
       totalItems = count;
       return Post.find()
-        .skip((currentPage - 1) * perPage) // skip first n posts
-        .limit(perPage); // limit to n posts
+        .skip((currentPage - 1) * perPage)
+        .limit(perPage);
     })
-    .then((posts) => {
-      res.status(200).json({ message: "Fetched posts.", posts: posts, totalItems: totalItems });
+    .then(posts => {
+      res.status(200).json({
+        message: 'Fetched posts successfully.',
+        posts: posts,
+        totalItems: totalItems
+      });
     })
-    .catch((err) => {
+    .catch(err => {
       if (!err.statusCode) {
         err.statusCode = 500;
       }
-      // pass error to next middleware
       next(err);
     });
 };
@@ -30,37 +36,43 @@ exports.getPosts = (req, res, next) => {
 exports.createPost = (req, res, next) => {
   const errors = validationResult(req);
   if (!errors.isEmpty()) {
-    // 422 - validation failed
-    const error = new Error("Validation failed, entered data is incorrect.");
+    const error = new Error('Validation failed, entered data is incorrect.');
     error.statusCode = 422;
     throw error;
   }
   if (!req.file) {
-    // 422 - validation failed
-    const error = new Error("No image provided.");
+    const error = new Error('No image provided.');
     error.statusCode = 422;
     throw error;
   }
   const imageUrl = req.file.path;
   const title = req.body.title;
   const content = req.body.content;
+  let creator;
   const post = new Post({
     title: title,
     content: content,
     imageUrl: imageUrl,
-    creator: { name: "Jedrzej" },
+    creator: req.userId
   });
-
   post
     .save()
-    .then((result) => {
-      console.log(result);
+    .then(result => {
+      return User.findById(req.userId);
+    })
+    .then(user => {
+      creator = user;
+      user.posts.push(post);
+      return user.save();
+    })
+    .then(result => {
       res.status(201).json({
-        message: "Post created successfully!",
-        post: result,
+        message: 'Post created successfully!',
+        post: post,
+        creator: { _id: creator._id, name: creator.name }
       });
     })
-    .catch((err) => {
+    .catch(err => {
       if (!err.statusCode) {
         err.statusCode = 500;
       }
@@ -71,20 +83,19 @@ exports.createPost = (req, res, next) => {
 exports.getPost = (req, res, next) => {
   const postId = req.params.postId;
   Post.findById(postId)
-    .then((post) => {
+    .then(post => {
       if (!post) {
-        // 404 - not found
-        const error = new Error("Could not find post.");
+        const error = new Error('Could not find post.');
         error.statusCode = 404;
         throw error;
       }
-      res.status(200).json({ message: "Post fetched.", post: post });
+      res.status(200).json({ message: 'Post fetched.', post: post });
     })
-    .catch((err) => {
+    .catch(err => {
       if (!err.statusCode) {
         err.statusCode = 500;
       }
-      next(err); // pass error to next middleware
+      next(err);
     });
 };
 
@@ -92,78 +103,89 @@ exports.updatePost = (req, res, next) => {
   const postId = req.params.postId;
   const errors = validationResult(req);
   if (!errors.isEmpty()) {
-    // 422 - validation failed
-    const error = new Error("Validation failed, entered data is incorrect.");
+    const error = new Error('Validation failed, entered data is incorrect.');
     error.statusCode = 422;
     throw error;
   }
   const title = req.body.title;
   const content = req.body.content;
   let imageUrl = req.body.image;
-
   if (req.file) {
     imageUrl = req.file.path;
   }
   if (!imageUrl) {
-    // 422 - validation failed
-    const error = new Error("No file picked.");
+    const error = new Error('No file picked.');
     error.statusCode = 422;
     throw error;
   }
-
   Post.findById(postId)
-    .then((post) => {
+    .then(post => {
       if (!post) {
-        // 404 - not found
-        const error = new Error("Could not find post.");
+        const error = new Error('Could not find post.');
         error.statusCode = 404;
+        throw error;
+      }
+      if (post.creator.toString() !== req.userId) {
+        const error = new Error('Not authorized!');
+        error.statusCode = 403;
         throw error;
       }
       if (imageUrl !== post.imageUrl) {
         clearImage(post.imageUrl);
       }
       post.title = title;
-      post.content = content;
       post.imageUrl = imageUrl;
+      post.content = content;
       return post.save();
     })
-    .then((result) => {
-      res.status(200).json({ message: "Post updated!", post: result });
+    .then(result => {
+      res.status(200).json({ message: 'Post updated!', post: result });
     })
-    .catch((err) => {
+    .catch(err => {
       if (!err.statusCode) {
         err.statusCode = 500;
       }
-      next(err); // pass error to next middleware
+      next(err);
     });
 };
 
 exports.deletePost = (req, res, next) => {
   const postId = req.params.postId;
   Post.findById(postId)
-    .then((post) => {
+    .then(post => {
       if (!post) {
-        // 404 - not found
-        const error = new Error("Could not find post.");
+        const error = new Error('Could not find post.');
         error.statusCode = 404;
         throw error;
       }
+      if (post.creator.toString() !== req.userId) {
+        const error = new Error('Not authorized!');
+        error.statusCode = 403;
+        throw error;
+      }
+      // Check logged in user
       clearImage(post.imageUrl);
       return Post.findByIdAndRemove(postId);
     })
-    .then((result) => {
-      console.log(result);
-      res.status(200).json({ message: "Post deleted!" });
+    .then(result => {
+      return User.findById(req.userId);
     })
-    .catch((err) => {
+    .then(user => {
+      user.posts.pull(postId);
+      return user.save();
+    })
+    .then(result => {
+      res.status(200).json({ message: 'Deleted post.' });
+    })
+    .catch(err => {
       if (!err.statusCode) {
         err.statusCode = 500;
       }
-      next(err); // pass error to next middleware
+      next(err);
     });
 };
 
-const clearImage = (filePath) => {
-  filePath = path.join(__dirname, "..", filePath);
-  fs.unlink(filePath, (err) => console.log(err));
+const clearImage = filePath => {
+  filePath = path.join(__dirname, '..', filePath);
+  fs.unlink(filePath, err => console.log(err));
 };
